requirements.txt
streamlit
pydantic
llama-cpp-python[metal]
huggingface-hub
pytest
transformers
torch
